requirements_to_pull.txt
beautifulsoup4==4.12.3
bs4==0.0.2
camply==0.32.8
Cartopy==0.23.0
certifi==2024.8.30
charset-normalizer==3.3.2
choix==0.3.5
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fake-useragent==1.4.0
fonttools==4.54.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
importlib_resources==6.4.5
kiwisolver==1.4.7
markdown-it-py==2.2.0
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.0.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pydantic==1.10.18
Pygments==2.18.0
pyparsing==3.1.4
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.2
ratelimit==2.2.1
requests==2.31.0
rich==13.3.5
rich-click==1.6.1
scipy==1.13.1
shapely==2.0.6
six==1.16.0
soupsieve==2.6
tenacity==8.2.3
trueskill==0.4.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
zipp==3.20.2
